+++ mex/backend/graph/connector.py
@@ -3,12 +3,16 @@ from collections.abc import Sequence
 from string import Template
 from typing import Annotated, Any, Literal, cast
 
+import backoff
+from backoff.types import Details as BackoffDetails
 from neo4j import Driver, GraphDatabase, NotificationDisabledCategory
+from neo4j.exceptions import DriverError
 from pydantic import Field
 
 from mex.backend.fields import SEARCHABLE_CLASSES, SEARCHABLE_FIELDS
+from mex.backend.graph.exceptions import InconsistentGraphError
 from mex.backend.graph.models import Result
-from mex.backend.graph.query import QueryBuilder
+from mex.backend.graph.query import Query, QueryBuilder
 from mex.backend.graph.transform import expand_references_in_search_result
 from mex.backend.settings import BackendSettings
 from mex.common.connector import BaseConnector
@@ -141,25 +145,47 @@ class GraphConnector(BaseConnector):
         """Close the connector's underlying requests session."""
         self.driver.close()
 
-    def commit(self, query: str, **parameters: Any) -> Result:
-        """Send and commit a single graph transaction."""
-        message = Template(query).safe_substitute(
-            {
-                k: json.dumps(v, ensure_ascii=False)
-                for k, v in (parameters or {}).items()
-            }
-        )
+    @staticmethod
+    def _should_giveup_commit(error: Exception) -> bool:
+        """When to give up on committing."""
+        return not cast(DriverError, error).is_retryable()
+
+    @staticmethod
+    def _on_commit_backoff(event: BackoffDetails) -> None:
+        """Re-connect to the graph database."""
+        self = cast(GraphConnector, event["args"][0])
         try:
-            with self.driver.session() as session:
-                result = Result(session.run(query, parameters))
-        except Exception as error:
-            logger.error("\n%s\n%s", message, error)
-            raise
-        if counters := result.get_update_counters():
-            logger.debug("\n%s\n%s", message, json.dumps(counters, indent=4))
+            self.close()
+        except DriverError as error:
+            logger.error("error closing before reconnect %s", error)
+        self.driver = self._init_driver()
+        self._check_connectivity_and_authentication()
+
+    @staticmethod
+    def _on_commit_giveup(event: BackoffDetails) -> None:
+        """Log the query when giving up on committing."""
+        query = cast(Query | str, event["args"][1])
+        kwargs = event["kwargs"]
+        settings = BackendSettings.get()
+        if settings.debug:
+            params = {k: json.dumps(v, ensure_ascii=False) for k, v in kwargs.items()}
+            message = f"\n{Template(str(query)).safe_substitute(params)}"
         else:
-            logger.debug("\n%s", message)
-        return result
+            message = f": {query!r}"
+        logger.error("error committing query%s", message)
+
+    @backoff.on_exception(
+        backoff.fibo,
+        DriverError,
+        giveup=_should_giveup_commit,
+        on_backoff=_on_commit_backoff,
+        on_giveup=_on_commit_giveup,
+        max_time=1000,
+    )
+    def commit(self, query: Query | str, **parameters: Any) -> Result:
+        """Send and commit a single graph transaction."""
+        with self.driver.session() as session:
+            return Result(session.run(str(query), parameters))
 
     def _fetch_extracted_or_rule_items(
         self,
@@ -468,13 +494,17 @@ class GraphConnector(BaseConnector):
             merged_label=ensure_prefix(model.stemType, "Merged"),
             ref_labels=ref_labels,
         )
-        return self.commit(
+        result = self.commit(
             query,
             **constraints,
             stable_target_id=stable_target_id,
             ref_identifiers=ref_identifiers,
             ref_positions=ref_positions,
         )
+        if len(result["edges"]) != len(ref_labels):
+            msg = "could not merge all edges"
+            raise InconsistentGraphError(msg)
+        return result
 
     def create_rule_set(
         self,
@@ -521,3 +551,16 @@ class GraphConnector(BaseConnector):
             self._merge_edges(model, model.stableTargetId, identifier=model.identifier)
 
         return [m.identifier for m in models]
+
+    def flush(self) -> None:
+        """Flush the database (only in debug mode)."""
+        settings = BackendSettings.get()
+        if settings.debug is True:
+            self.driver.execute_query("MATCH (n) DETACH DELETE n;")
+            for row in self.driver.execute_query("SHOW ALL CONSTRAINTS;").records:
+                self.driver.execute_query(f"DROP CONSTRAINT {row['name']};")
+            for row in self.driver.execute_query("SHOW ALL INDEXES;").records:
+                self.driver.execute_query(f"DROP INDEX {row['name']};")
+        else:
+            msg = "database flush was attempted outside of debug mode"
+            raise MExError(msg)